ResourceKinds and Self Affecting Relationships

+++ Assets/Scripts/BubbleRelationshipDefinition.cs
@@ -5,5 +5,5 @@ public abstract class BubbleRelationshipDefinition : ScriptableObject
     public BubbleDefinition DefinitionA;
     public BubbleDefinition DefinitionB;
 
-    public abstract void ExecuteRelationship(Bubble bubbleA, Bubble bubbleB, float time);
+    public abstract void ExecuteRelationship(Bubble bubbleA, Bubble bubbleB, double time);
 }
